разделяет проект на модули
